Move CFRChange navigation elements into fr_notices
They still need to be rewritten to remove the duplication
