+++ regulations/tests/views_preamble_tests.py
@@ -4,7 +4,7 @@ from mock import patch
 from unittest import TestCase
 from datetime import date, timedelta
 
-from nose.tools import assert_equal, assert_is_none
+from nose.tools import assert_equal
 from django.http import Http404
 from django.test import RequestFactory, override_settings
 
@@ -38,10 +38,10 @@ class PreambleViewTests(TestCase):
         self.assertIsNone(fn(root, ['1', 'c', 'r']))
         self.assertIsNone(fn(root, ['1', 'c', 'i', 'r']))
 
-    @patch('regulations.views.preamble.CFRChangeToC')
+    @patch('fr_notices.navigation.CFRChangeBuilder')
     @patch('regulations.generator.generator.api_reader')
     @patch('regulations.views.preamble.ApiReader')
-    def test_get_integration(self, ApiReader, api_reader, CFRChangeToC):
+    def test_get_integration(self, ApiReader, api_reader, CFRChangeBuilder):
         """Verify that the contexts are built correctly before being sent to
         the template. AJAX/partial=true requests should only get the inner
         context (i.e. no UI-related context)"""
@@ -140,10 +140,11 @@ class PreambleViewTests(TestCase):
         _, meta, _ = preamble.notice_data('1')
         assert_equal(meta['comment_state'], CommentState.CLOSED)
 
-    @patch('regulations.views.preamble.CFRChangeToC')
+    @patch('fr_notices.navigation.CFRChangeBuilder')
     @patch('regulations.generator.generator.api_reader')
     @patch('regulations.views.preamble.ApiReader')
-    def test_get_top_level_redirect(self, ApiReader, api_reader, CFRChangeToC):
+    def test_get_top_level_redirect(self, ApiReader, api_reader,
+                                    CFRChangeBuilder):
         ApiReader.return_value.preamble.return_value = self._mock_preamble
         api_reader.ApiReader.return_value.layer.return_value = {
             '1-c-x': ['something']
@@ -240,133 +241,3 @@ class CFRChangesViewTests(TestCase):
         self.assertIn('111-43', subpart_info.urls[1])
         self.assertIn('111-44', subpart_info.urls[2])
         self.assertIn('111-45', subpart_info.urls[3])
-
-
-class CFRChangeToCTests(TestCase):
-    @patch('regulations.views.preamble.fetch_toc')
-    @patch('regulations.views.preamble.utils.regulation_meta')
-    def test_add_amendment(self, fetch_meta, fetch_toc):
-        """Add amendments for two different CFR parts. Verify that the table
-        of contents contains only the changed data. Also validate that changes
-        to subparts do not include a ToC entry"""
-        version_info = {'111': {'left': 'v1', 'right': 'v2'},
-                        '222': {'left': 'vold', 'right': 'vnew'}}
-        builder = preamble.CFRChangeToC('docdoc', version_info)
-        fetch_toc.return_value = [
-            dict(index=['111', '1'], title='Section 1'),
-            dict(index=['111', '1', 'a'], title='1 a'),
-            dict(index=['111', '2'], title='Section 2'),
-            dict(index=['111', '3'], title='Section 3')]
-        fetch_meta.return_value = dict(cfr_title_number='99',
-                                       statutory_name='Some title for reg 111')
-        builder.add_amendment(dict(cfr_part='111', instruction='1. inst1',
-                                   authority='auth1'))
-        # subpart change -- doesn't affect ToC
-        builder.add_amendment(dict(cfr_part='111', instruction='2. inst2',
-                                   changes=[['111-Subpart-A', []]]))
-        builder.add_amendment(dict(cfr_part='111', instruction='2. inst2',
-                                   # The second element of each pair would be
-                                   # non-empty in realistic scenarios
-                                   changes=[['111-1', []], ['111-3-b', []]]))
-
-        fetch_toc.return_value = [dict(index=['222', '4'], title='Section 4')]
-        fetch_meta.return_value = dict(cfr_title_number='99',
-                                       statutory_name='Some title for reg 222')
-        # only authority change
-        builder.add_amendment(dict(cfr_part='222', instruction='3. inst3',
-                                   authority='auth2'))
-
-        self.assertEqual(builder.toc, [
-            preamble.ToCPart(
-                title='99', part='111', name='Some title for reg 111',
-                authority_url='/preamble/docdoc/cfr_changes/111',
-                sections=[
-                    preamble.ToCSect(section='1', title='Section 1',
-                                     url='/preamble/docdoc/cfr_changes/111-1',
-                                     full_id='docdoc-cfr-111-1', part='111'),
-                    preamble.ToCSect(section='3', title='Section 3',
-                                     url='/preamble/docdoc/cfr_changes/111-3',
-                                     full_id='docdoc-cfr-111-3', part='111')
-                ]),
-            preamble.ToCPart(
-                title='99', part='222', name='Some title for reg 222',
-                authority_url='/preamble/docdoc/cfr_changes/222',
-                sections=[])])
-
-    def test_navigation(self):
-        toc = [
-            preamble.ToCPart(
-                part='478',
-                title='27',
-                name='Commerce',
-                authority_url='',
-                sections=[
-                    preamble.ToCSect(
-                        full_id='2016_02749-cfr-478-99',
-                        url='/preamble/2016_02749/cfr_changes/478-99',
-                        title=u'§ 478.99 Certain prohibited',
-                        part='478', section='99',
-                    ),
-                    preamble.ToCSect(
-                        full_id='2016_02749-cfr-478-120',
-                        url='/preamble/2016_02749/cfr_changes/478-120',
-                        title=u'§ 478.120 Firearms',
-                        part='478', section='120',
-                    ),
-                ],
-            )
-        ]
-
-        nav = preamble.section_navigation([], toc, part='478', section='99')
-        assert_equal(nav['next'].section_id, '2016_02749-cfr-478-120')
-        assert_equal(nav['previous'].markup_prefix, '27 CFR 478')
-
-        nav = preamble.section_navigation([], toc, part='478', section='120')
-        assert_equal(nav['previous'].section_id, '2016_02749-cfr-478-99')
-        assert_is_none(nav['next'])
-
-    def test_get_toc_position(self):
-        toc = [
-            preamble.ToCPart(
-                part='478',
-                title='27',
-                name='Commerce',
-                authority_url='',
-                sections=[
-                    preamble.ToCSect(
-                        full_id='2016_02749-cfr-478-99',
-                        url='/preamble/2016_02749/cfr_changes/478-99',
-                        title=u'§ 478.99 Certain prohibited',
-                        part='478', section='99',
-                    ),
-                    preamble.ToCSect(
-                        full_id='2016_02749-cfr-478-120',
-                        url='/preamble/2016_02749/cfr_changes/478-120',
-                        title=u'§ 478.120 Firearms',
-                        part='478', section='120',
-                    ),
-                ],
-            ),
-            preamble.ToCPart(
-                part='521',
-                title='27',
-                name='Commerce',
-                authority_url='',
-                sections=[
-                    preamble.ToCSect(
-                        full_id='2016_02749-cfr-521-18',
-                        url='/preamble/2016_02749/cfr_changes/521-18',
-                        title=u'§ 521.18 Something else',
-                        part='521', section='18',
-                    )
-                ],
-            )]
-
-        self.assertEqual(0, preamble.get_toc_position(toc,
-                         **{'part': '478', 'section': '99'}))
-        self.assertEqual(1, preamble.get_toc_position(toc,
-                         **{'part': '478', 'section': '120'}))
-        self.assertEqual(2, preamble.get_toc_position(toc,
-                         **{'part': '521', 'section': '18'}))
-        self.assertIsNone(preamble.get_toc_position(toc,
-                          **{'part': 'non-existent', 'section': '18'}))
